reverted the mongo primitives
mongo, for some weird reason i dont know, does not like Float, so it
translates them into Doubles. Thats why Pitch and Yaw were Doubles in
the first place.

+++ src/main/java/io/github/apfelcreme/SupportTickets/Bukkit/Bungee/BungeeMessageListener.java
@@ -43,7 +43,7 @@ public class BungeeMessageListener implements PluginMessageListener {
             UUID uuid = UUID.fromString(in.readUTF());
             Location location = new Location(SupportTickets.getInstance().getServer().getWorld(in.readUTF()),
                     in.readDouble(), in.readDouble(), in.readDouble(),
-                    in.readFloat(), in.readFloat());
+                    (float) in.readDouble(), (float) in.readDouble());
             warp(uuid, location);
         } else if (subChannel.equals("POSITIONREQUEST")) {
             UUID uuid = UUID.fromString(in.readUTF());
@@ -68,13 +68,13 @@ public class BungeeMessageListener implements PluginMessageListener {
             if (player != null) {
                 out.writeUTF("POSITIONANSWER");
                 out.writeUTF(uuid.toString());
-                out.writeUTF(SupportTickets.getInstance().getServer().getIp()+":"+SupportTickets.getInstance().getServer().getPort());
+                out.writeUTF(SupportTickets.getInstance().getServer().getIp() + ":" + SupportTickets.getInstance().getServer().getPort());
                 out.writeUTF(player.getWorld().getName());
                 out.writeDouble(player.getLocation().getX());
                 out.writeDouble(player.getLocation().getY());
                 out.writeDouble(player.getLocation().getZ());
-                out.writeFloat(player.getLocation().getYaw());
-                out.writeFloat(player.getLocation().getPitch());
+                out.writeDouble((double) player.getLocation().getYaw());
+                out.writeDouble((double) player.getLocation().getPitch());
                 out.writeUTF(message);
                 player.sendPluginMessage(SupportTickets.getInstance(), "SupportTickets", b.toByteArray());
                 out.close();
